references/04_date_processing/requirements_ai_extractor.txt
numpy>=1.21.0
openpyxl>=3.0.0
openai>=1.0.0
pymongo>=4.0.0
rapidfuzz>=3.0.0
python-dateutil>=2.8.0 